+++ src/components/hotel/Floor.tsx
