+++ src/components/table-container/TableCell.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import { hot } from "react-hot-loader";
 
-import { useAppSelector } from "../redux/hooks";
+import { useAppSelector } from "../../redux/hooks";
+import * as table from "../../redux/tableSlice";
 
 import Tile from "./Tile";
 
 import "./TableCell.css";
-import { TileData } from "../redux/tableSlice";
 
 type Props = {
   x: number,
@@ -31,7 +31,7 @@ function TableCell(props: Props): JSX.Element {
   return <div className="table-cell">{tile}</div>;
 }
 
-function useTileDataAt(x: number, y: number): TileData | undefined {
+function useTileDataAt(x: number, y: number): table.TileData | undefined {
   return useAppSelector(state => {
     const occupations = state.table.occupations;
     const occupationsForRoom = occupations[y];
